requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.4
django-cron==0.6.0
django-extensions==3.2.3
django-import-export==3.3.7
django-jalali==6.2.1
et-xmlfile==1.1.0
frozenlist==1.4.1
idna==3.7
jalali_core==1.0.0
jdatetime==5.0.0
MarkupPy==1.14
multidict==6.0.5
odfpy==1.4.1
openpyxl==3.1.2
PyJWT==2.8.0
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.5.0
tablib==3.5.0
twilio==9.0.5
tzdata==2024.1
urllib3==2.2.1
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
